Spear Web 2.0
This version of the app uses SQLAlchemy to make the SQL queries and manage all of the databased information. In addition forms are dealt with using a built in forms class. Currently there is no graphical view of the printing out of files in the directory, but that will be added soon.